# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Research Model Builder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{id}/papers/{paperId}/extract:
    post:
      operationId: extractVariables
      tags: [variables]
      summary: Extract research variables from a paper using AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: paperId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Extracted variables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Variable"
  /sessions/{id}/variables:
    get:
      operationId: listSessionVariables
      tags: [variables]
      summary: List all extracted variables for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Variables list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Variable"
...
  /sessions/{id}/variables/{variableId}:
    patch:
      operationId: updateSessionVariable
      tags: [variables]
      summary: Manually rename, retype, or rewrite the definition of a single variable. Other downstream views (graph, models, live model) read on demand and will reflect the change.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: variableId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: [independent, mediator, moderator, dependent]
                definition:
                  type: string
      responses:
        "200":
          description: Updated variable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Variable"
        "400":
          description: No supported fields to update / invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Variable not found in this session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSessionVariable
      tags: [variables]
      summary: Delete a single variable from the session.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: variableId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "404":
          description: Variable not found in this session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sessions/{id}/variable-graph:
    get:
      operationId: getVariableGraph
      tags: [variables]
      summary: Get variable relationship graph data for visualization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Graph nodes and edges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VariableGraph"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Variable:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        paperId:
          type: integer
        paperTitle:
          type: string
        paperAuthors:
          type: array
          items:
            type: string
        paperYear:
          type: ["integer", "null"]
        name:
          type: string
        type:
          type: string
          enum: [independent, mediator, moderator, dependent]
        definition:
          type: string
        citationText:
          type: string
        canonicalConstructId:
          type: ["string", "null"]
          description: Lower-case canonical construct id (e.g. "trust", "purchase intention"). Variables across papers that map to the same construct share this id.
        constructLayer:
          type: ["string", "null"]
          enum: [stimulus, cognitive, affective, intention, behavior, null]
          description: Standard psychology pipeline layer for the variable.
        createdAt:
          type: string
      required:
        - id
        - sessionId
        - paperId
        - paperTitle
        - paperAuthors
        - name
        - type
        - definition
        - citationText
        - createdAt
...
    VariableGraph:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
      required:
        - nodes
        - edges
    GraphNode:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
        paperCount:
          type: integer
      required:
        - id
        - label
        - type
        - paperCount
    GraphEdge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        paperId:
          type: integer
        paperTitle:
          type: string
        relationship:
          type: string
          enum: [positive, negative, moderates, mediates]
          description: Kind of directional relationship the paper states between source and target.
        statement:
          type: string
          description: Verbatim sentence from the paper supporting this edge.
      required:
        - source
        - target
        - paperId
        - paperTitle
        - relationship
        - statement
...